add waitingList remove check dialog

+++ app/src/main/java/kr/ac/tukorea/waiter/Waiting_List_Page.kt
@@ -1,15 +1,17 @@
 package kr.ac.tukorea.waiter
 
 import android.app.Activity
+import android.app.AlertDialog
+import android.app.Dialog
 import android.content.ContentValues.TAG
+import android.content.Context
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import android.view.Menu
 import android.view.View
-import android.widget.AdapterView
-import android.widget.ImageButton
-import android.widget.Toast
+import android.view.WindowManager
+import android.widget.*
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ktx.firestore
@@ -17,16 +19,18 @@ import com.google.firebase.firestore.ktx.toObject
 import com.google.firebase.ktx.Firebase
 import kotlinx.android.synthetic.main.activity_sign_in.*
 import kotlinx.android.synthetic.main.activity_waiting_list_page.*
-import kr.ac.tukorea.waiter.databinding.ActivityWaitingListPageBinding
+import kr.ac.tukorea.waiter.databinding.WaitingListPageLayoutBinding
 import java.text.FieldPosition
 
 
 class Waiting_List_Page : AppCompatActivity() {
 
-    private var Wbinding: ActivityWaitingListPageBinding? = null
-    private val binding get() = Wbinding!!
+//    private var Wbinding: DialogmenuBinding? = null
+//    private val binding get() = Wbinding!!
 
     var db: FirebaseFirestore = Firebase.firestore
+
+    lateinit var newbtn : Button
 
     data class UserInfo(
         val index: String? = null,
@@ -34,6 +38,7 @@ class Waiting_List_Page : AppCompatActivity() {
         val phone: String? = null,
         val customeNum: Int = 0
     )
+
 
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {  //메뉴
@@ -52,6 +57,8 @@ class Waiting_List_Page : AppCompatActivity() {
 
         val reserveInfo = db.collection("rest_Info").document("abc")
             .collection("reservation")
+
+
 
         reserveInfo.get().addOnSuccessListener { result ->
 
@@ -87,21 +94,43 @@ class Waiting_List_Page : AppCompatActivity() {
 
                 val indexpo = (position+1).toString()
 
-                val remove = db.collection("rest_Info").document("abc")
-                    .collection("reservation").document("${indexpo}")
-                remove.delete().addOnSuccessListener {
-                    Log.d("Testdata","succes")
-                }
+                AlertDialog.Builder(this)  // 다이어로그 출력
+                    .setView(R.layout.dialogmenu)
+                    .show()
+                    .also{ alertDialog ->
+                        if(alertDialog == null){
+                            return@also
+                        }
+//                        Wbinding = DialogmenuBinding.inflate(layoutInflater)
+//                        setContentView(binding.root)  //바인딩 연결 시간 남으면
+                        val button1 = alertDialog.findViewById<Button>(R.id.diabtn1)
+                        val button2 = alertDialog.findViewById<Button>(R.id.diabtn2)
+                        val infotext = alertDialog.findViewById<TextView>(R.id.infoText)
+
+                        infotext.setText("${(position+1)}번째 리스트를 삭제")
+
+                        button1?.setOnClickListener {
+                            alertDialog.dismiss()
+                            Log.d("Testdata","확인")
+
+                            val remove = db.collection("rest_Info").document("abc")   //DB삭제
+                                .collection("reservation").document("${indexpo}")
+                            remove.delete().addOnSuccessListener {
+                                Log.d("Testdata","succes")
+                            }
+                        }
+
+                        button2?.setOnClickListener {
+                            alertDialog.dismiss()
+                            Log.d("Testdata","취소")
+                        }
+
+                    }
+
+
             }
 
         }
-
-        var checked = 0
-        //items.remove
-
-        Wbinding = ActivityWaitingListPageBinding.inflate(layoutInflater)
-        setContentView(binding.root)  //바인딩 연결
-
 
     }
 
@@ -110,6 +139,7 @@ class Waiting_List_Page : AppCompatActivity() {
 
         setContentView(R.layout.activity_waiting_list_page)
         var reintent = intent
+
 
 
     }
